index.js 내용 SparkPolicy로 캡슐화, index.ejs에 applist및userlist jquery로 처리, 필요없어진 파일 및 코드 제거
