+++ spark_node/views/index.ejs
@@ -7,23 +7,14 @@
   </head>
   <body>
     <p>
-    <form id='' action='spk'>
+    <form id='list' action='/makeList'>
       <select name='spkfile' size='6' id='select_APP'>
-        <option value='' selected>--select Spark APP--</option>
-	<% for (var i =0 ; i<applist.length ; i++){ %>
-	<option value='<%= applist[i] %>'><%= applist[i] %></option>
-	<% } %>
       </select>
 
       <select name='username' size='6' id='select_user' class='select_user'>
-        <option value='' selected>--select user folder--</option>
-	<% for (var i =1 ; i<userlist.length ; i++){ %>
-	<option value='<%= userlist[i] %>' id='select_user<%=i%>' class='select_user<%=i%>'><%= userlist[i] %></option>
-	<% } %>
       </select>
 
       <select name='textfile' size='6' id='select_data'>
-	<option value='' selected>--select data file--</option>
       </select>
     </p>
 
@@ -31,15 +22,42 @@
 
     </form>
 
-      <button class="submit">전송하기</button>
+    <button class="submit">spark-submit</button>
 
-      <div class="result">결과</div>
+    <div class="result">결과</div>
+
+    <p>upload spark app</p>
+
+    <form id="FILE_FORM" method='post' enctype='multipart/form-data' action=''>
+      <input type='file' id='FILE_TAG' name='FILE_TAG'>
+    </form>
+    <button id='file-upload'>파일전송</button>
+
+<script>
+  $('#file-upload').click( function() {
+//    var form = $('#FILE_FORM')[0];
+//    var formData = new Formdata(form);
+//    formData.append('fileObj', $('#FILE_TAG')[0].file[0]);
+
+    $.ajax({
+      url: '',
+	processData: false,
+	contentType: multipart/form-data,
+	data: $('#FILE_TAG')[0].file[0],
+	type: 'POST',
+	sucess: function(result){
+	  alert(result+'업로드 성공');
+	}
+    });
+  }
+</script>
 
 
-
+<!--//////////////////////////////////////////////-->
+<!--
 <button id='startSparkBtn'>spark-submit</button>
 <h4 id="resultSparkTxt">result : </h4>
-<!--<form action="/yarnAll" method="GET">
+<form action="/yarnAll" method="GET">
 	<button type=submit>button1</button>
  </form> -->
 <!--<form action="/uploads" method="post" enctype="multipart/form-data">
@@ -152,7 +170,7 @@ function allRefreshState() {
 $('.submit').click( function() {
       $('.result').html('처리중');
       $.ajax({
-        url: '/spk',
+        url: '/sparkSubmit',
         method: 'post',
         data: {
 	'APP' : $('#select_APP option:selected').val(),
@@ -168,12 +186,9 @@ $('.submit').click( function() {
         }
       });
     });
-  </script>
-
-  <script>
     $('#select_user').click( function() {
       $.ajax({
-        url: '/spk',
+        url: '/makeList',
         method: 'post',
         data: {
 	'user' : $('#select_user option:selected').val()
@@ -188,6 +203,26 @@ $('.submit').click( function() {
 	    }
           }
         }
+      });
+    });
+    $(document).ready(function(){
+      $.ajax({
+        url: '/makeList',
+        method: 'post',
+        data: {
+	'user' : $('#select_user option:selected').val()
+	},
+        dataType: 'json',
+        success: function(result) {
+          if(result){
+	    for(var i=1 ; i<result.applist.length ; i++){
+              $('#select_APP').append('<option value='+result.applist[i]+'>'+result.applist[i]+'</option>');
+	    }
+	    for(var i=1 ; i<result.userlist.length ; i++){
+              $('#select_user').append('<option value='+result.userlist[i]+'>'+result.userlist[i]+'</option>');
+	    }
+	  }
+	}
       });
     });
 	</script>
