--- spark_node/views/spk.ejs
@@ -1,94 +1,0 @@
-<!DOCTYPE html>                                            
-
-<html>
-
-  <head>
-
-    <title>Spark</title>
-
-    <link rel='stylesheet' href='/stylesheets/style.css' />
-
-    <script src="//code.jquery.com/jquery-3.2.1.min.js"></script>
-
-  </head>
-
-  <body>
-
-    <h1>Spark</h1>
-
-    <p>
-    <form id='' action='spk'>
-      <select name='spkfile' size='6' id='select_APP'>
-        <option value='' selected>--select Spark APP--</option>
-	<% for (var i =0 ; i<applist.length ; i++){ %>
-	<option value='<%= applist[i] %>'><%= applist[i] %></option>
-	<% } %>
-      </select>
-
-      <select name='username' size='6' id='select_user' class='select_user'>
-        <option value='' selected>--select user folder--</option>
-	<% for (var i =1 ; i<userlist.length ; i++){ %>
-	<option value='<%= userlist[i] %>' id='select_user<%=i%>' class='select_user<%=i%>'><%= userlist[i] %></option>
-	<% } %>
-      </select>
-
-      <select name='textfile' size='6' id='select_data'>
-	<option value='' selected>--select data file--</option>
-      </select>
-    </p>
-
-      <input id='paramater' name='paramater' type='text' placeholder='input paramater'>
-
-    </form>
-
-      <button class="submit">전송하기</button>
-
-      <div class="result">결과</div>
-
-  <script>
-    $('.submit').click( function() {
-      $('.result').html('처리중');
-      $.ajax({
-        url: '/spk',
-        method: 'post',
-        data: {
-	'APP' : $('#select_APP option:selected').val(),
-	'user' : $('#select_user option:selected').val(),
-	'data' : $('#select_data option:selected').val(),
-	'paramater' : $('#paramater').val()
-	},
-        dataType: 'json',
-        success: function(result) {
-          if(result){
-            $('.result').html(result.result);
-          }
-        }
-      });
-    });
-  </script>
-
-  <script>
-    $('#select_user').click( function() {
-      $.ajax({
-        url: '/spk',
-        method: 'post',
-        data: {
-	'user' : $('#select_user option:selected').val()
-	},
-        dataType: 'json',
-        success: function(result) {
-          if(result){
-
-	    $('#select_data').empty();
-	    for(var i=1 ; i<result.datalist.length ; i++){
-              $('#select_data').append('<option value='+result.datalist[i]+'>'+result.datalist[i]+'</option>');
-	    }
-          }
-        }
-      });
-    });
-  </script>
-
-  </body>
-
-</html>
