--- spark_node/views/uploadApp.ejs
@@ -1,11 +1,0 @@
-<doctype html>
-<html>
-   <head>
-   </head>
-   <body>
-	<p>upload app file</p>
-	<form action='upload' method='post' enctype="multipart/form-data">
-		<input type='file' value='choose' name='userfile'>
-		<input type='submit' value='submit'>
-   </body>
-</html>
